se agrego una imagen de fondo, se corrigio el carrousel imgenes responsitivas y desplazamiento
