+++ css/main.css
@@ -48,6 +48,20 @@ body {
     align-items: center;
 }
 
+.background {
+    position: fixed;
+    top: 10em;
+    left: 0;
+    width: 100%;
+    height: 100%;
+    background-image: url('../src/background.jpg');
+    background-size: cover;
+    background-attachment: fixed;
+    background-position: center;
+    background-blend-mode: multiply;
+    z-index: -1;
+}
+
 header {
     width: 100%;
     height: 10em;
@@ -247,6 +261,8 @@ nav .container .menu-toggle .bar {
     left: 0;
     z-index: -1;
     height: 100%;
+    aspect-ratio: 16 / 9; 
+    object-fit: cover; 
 }
 
 .carousel-control-prev,
