+++ index.html
@@ -38,7 +38,8 @@
     </ul>
   </div>
 </nav>
-<main>
+<div class="background"></div>
+<main >
   <div class="container">
     <section id="novedades" class="bungee-regular">
       <h2>Recien llegados</h2>
